Entirely new project
In my naïvety as a young developer this project was originally copied
from another and the package names remained the same.  Once Proguard
was introduced and changes to the foundational .jar libraries were
accessed duplicate files were halting the build process.  Builds to
production were fine but testing was not possible.  As the upgrade to
this application is going to integrate a large amount of both unit,
integration, and UI testing I figured it best for a fresh start.